+++ Sort/BubbleSort/bubble_sort.h
@@ -3,6 +3,6 @@
 
 #include <cstddef>
 
-void bubbleSort(int arr[], const size_t size);
+void bubbleSort(int arr[], const int size);
 
 #endif // BUBBLE_SORT_H
